Refactor local functions into service file
